# Conflicts:
#	Source/Clients/DotNET/EventStore.cs
#	Source/Kernel/Contracts/Observation/IObservers.cs
#	Source/Kernel/Contracts/Observation/ObserverInformation.cs
#	Source/Kernel/Core.Specs/Services/Observation/for_ObserverInformationConverters/when_converting_to_contract/and_observer_is_not_replayable.cs
#	Source/Kernel/Core.Specs/Services/Observation/for_ObserverInformationConverters/when_converting_to_contract/and_observer_is_replayable.cs
#	Source/Kernel/Core.Specs/Services/Observation/for_Observers/when_getting_replayable_observers_for_event_types/and_no_observer_states_match.cs
#	Source/Kernel/Core.Specs/Services/Observation/for_Observers/when_getting_replayable_observers_for_event_types/and_there_are_matching_observers.cs
#	Source/Kernel/Core/Services/Observation/ObserverInformationConverters.cs
#	Source/Kernel/Core/Services/Observation/Observers.cs

+++ Source/Kernel/Storage.Sql/EventStores/Namespaces/UniqueConstraints/UniqueConstraintIndexEntryConverter.cs
@@ -19,7 +19,7 @@ public static class UniqueConstraintIndexEntryConverter
         new(
             entry.EventSourceId,
             entry.Value,
-            (EventSequenceNumber)entry.SequenceNumber);
+            (EventSequenceNumber)(ulong)entry.SequenceNumber);
 
     /// <summary>
     /// Convert from <see cref="UniqueConstraintIndex"/> to <see cref="UniqueConstraintIndexEntry"/>.
